remove common bundle and create widget specific bundles
