--- src/common/nls/common.ts
@@ -1,50 +1,0 @@
-const locales = {
-	zh: () => import('./zh-CN/common'),
-	'zh-TW': () => import('./zh-TW/common')
-};
-
-const messages = {
-	sunShort: 'Sun',
-	monShort: 'Mon',
-	tueShort: 'Tue',
-	wedShort: 'Wed',
-	thuShort: 'Thu',
-	friShort: 'Fri',
-	satShort: 'Sat',
-	sunday: 'Sunday',
-	monday: 'Monday',
-	tuesday: 'Tuesday',
-	wednesday: 'Wednesday',
-	thursday: 'Thursday',
-	friday: 'Friday',
-	saturday: 'Saturday',
-	janShort: 'Jan',
-	febShort: 'Feb',
-	marShort: 'Mar',
-	aprShort: 'Apr',
-	mayShort: 'May',
-	junShort: 'Jun',
-	julShort: 'Jul',
-	augShort: 'Aug',
-	sepShort: 'Sep',
-	octShort: 'Oct',
-	novShort: 'Nov',
-	decShort: 'Dec',
-	january: 'January',
-	february: 'February',
-	march: 'March',
-	april: 'April',
-	may: 'May',
-	june: 'June',
-	july: 'July',
-	august: 'August',
-	september: 'September',
-	october: 'October',
-	november: 'November',
-	december: 'December',
-	clear: 'clear',
-	close: 'close',
-	open: 'open'
-};
-
-export default { locales, messages };
