+++ src/dialog/nls/zh-CN/Dialog.tsx
@@ -1,0 +1,3 @@
+export default {
+	close: '关闭'
+};
