+++ src/dialog/nls/zh-TW/Dialog.tsx
@@ -1,0 +1,3 @@
+export default {
+	close: '關閉'
+};
