+++ src/slide-pane/nls/SlidePane.tsx
@@ -1,0 +1,10 @@
+const locales = {
+	zh: () => import('./zh-CN/SlidePane'),
+	'zh-TW': () => import('./zh-TW/SlidePane')
+};
+
+const messages = {
+	close: 'close'
+};
+
+export default { locales, messages };
