+++ src/slide-pane/nls/zh-CN/SlidePane.tsx
@@ -1,0 +1,3 @@
+export default {
+	close: '关闭'
+};
