+++ src/slide-pane/nls/zh-TW/SlidePane.tsx
@@ -1,0 +1,3 @@
+export default {
+	close: '關閉'
+};
