+++ src/tab-container/nls/TabContainer.tsx
@@ -1,0 +1,10 @@
+const locales = {
+	zh: () => import('./zh-CN/TabContainer'),
+	'zh-TW': () => import('./zh-TW/TabContainer')
+};
+
+const messages = {
+	close: 'close'
+};
+
+export default { locales, messages };
