+++ src/tab-container/nls/zh-CN/TabContainer.tsx
@@ -1,0 +1,3 @@
+export default {
+	close: '关闭'
+};
