+++ src/tab-container/nls/zh-TW/TabContainer.tsx
@@ -1,0 +1,3 @@
+export default {
+	close: '關閉'
+};
